Bug: #92, #297 - Fixed the image jitter problem when an image is zoomed or panned.  Zooming and panning is now smooth. This was done by replacing ints with floats in certain places. - Removed the exception eating in the fixed-point bilinear interpolation code (#297) - Removed the slow bilinear interpolation code altogether - Removed the nearest neighbour interpolation code, since it still causes jitter, and no one uses it now anyway. Should reimplement later as a C dll and re-add the NearestNeighbour option to the InterpolationMethod enum. - Simplified the interpolator code so that only static functions are called - Tested with all standard sample images. Zoom, pan, flip, rotation and resize all seem to work okay.

+++ ImageViewer/Rendering/BilinearInterpolation/BilinearInterpolation.h
@@ -25,19 +25,19 @@ extern "C"
 			BOOL isRGB,
 			BOOL isPlanar,
 
-			int srcRegionRectLeft,
-            int srcRegionRectTop,
-            int srcRegionRectRight,
-            int srcRegionRectBottom,
+			float srcRegionRectLeft,
+            float srcRegionRectTop,
+            float srcRegionRectRight,
+            float srcRegionRectBottom,
 			
             BYTE* pDstPixelData,
             unsigned int dstWidth,
             unsigned int dstBytesPerPixel,
 
-			int dstRegionRectLeft,
-            int dstRegionRectTop,
-            int dstRegionRectRight,
-            int dstRegionRectBottom,
+			float dstRegionRectLeft,
+            float dstRegionRectTop,
+            float dstRegionRectRight,
+            float dstRegionRectBottom,
 
 			BOOL swapXY,
             int* pLutData
